% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\title{Alpha-hull computation}
\usage{
complement(x, y = NULL, alpha)
}
\arguments{
\item{x}{coordinates of the sites}

\item{y}{coordinates of the sites}

\item{alpha}{positive parameter}
}
\value{
A matrix containing information about balls and halfplanes constituting the complement of the 
alpha hull
}
\description{
Computes the complement of an alpha hull
}
\examples{
x = runif(10)
y = runif(10)
alpha = 2
ahull.compl = complement(x, y, alpha)

}
