% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delvor.R
\name{delvor}
\alias{delvor}
\title{Voronoi diagram/Delanuay tesselation}
\usage{
delvor(x, y = NULL)
}
\arguments{
\item{x}{coordinates of the sites}

\item{y}{coordinates of the sites}
}
\value{
An invisible object of class "del.vor", a list, with the following components:
\describe{
   \item{mesh}{a mesh describing the diagram}
   \item{x}{nx2 matrix containing the coordinates of the sites}
   \item{tri.obj}{a "fake" tri object describing the triangulation}
}
}
\description{
Computing Voronoi diagram and Delanuay tesselation for the specified set of points
}
\examples{
x = runif(10)
y = runif(10)
vor.del = delvor(x, y)

}
