% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inahull.R
\name{inahull}
\alias{inahull}
\title{Determines if the provided points fall inside the \eqn{\alpha}-hull or not.}
\usage{
inahull(ahull.obj, x, y = NULL, alpha)
}
\arguments{
\item{x}{x coordinates of the points to be checked. Alternatively, a matrix with two columns containing both 
coordinates.}

\item{y}{y coordinates of the points to be checked.}

\item{ahull.onj}{an object class "ahull" like the one returned by the function \code{\link{ahull}}}
}
\value{
A logical vector of the same length of the number of points and such that the i-th element denotes 
if the i-th point is in the \eqn{\alpha}-hull or not.
}
\description{
Determines if the provided points fall inside the \eqn{\alpha}-hull or not by calling an external C++
function.
}
\examples{
x = runif(10)
y = runif(10)
inahull(ahull.obj, x, y)

}
\seealso{
\code{\link{ahull}}, \code{\link{complement}}
}
