% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.hall.R
\name{my.ahull}
\alias{my.ahull}
\title{Alpha-hull computation}
\usage{
my.ahull(x, y = NULL, alpha)
}
\arguments{
\item{x}{coordinates of the sites}

\item{y}{coordinates of the sites}
}
\value{
A list with the following components
}
\description{
Computes the alpha-hull for the given set of points
}
\examples{
x = runif(10)
y = runif(10)
a.hull = my.ahull(x, y)

}
