% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.ashape.R
\name{my.ashape}
\alias{my.ashape}
\title{Alpha-shape computation}
\usage{
my.ashape(x, y = NULL, alpha)
}
\arguments{
\item{x}{coordinates of the sites}

\item{y}{coordinates of the sites}
}
\value{
A list with the following components
}
\description{
Computes the alpha-shape for the given set of points
}
\examples{
x = runif(10)
y = runif(10)
a.shape = my.ashape(x, y, alpha = 0.8)

}
