% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ahull.not.R
\name{plot.ahull.not}
\alias{plot.ahull.not}
\title{\eqn{\alpha}-hull plot}
\usage{
\method{plot}{ahull.not}(ahull.obj, add = FALSE, do.shape = FALSE,
  wlines = c("none", "both", "del", "vor"), wpoints = TRUE,
  number = FALSE, col = NULL, xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{ahull.obj}{object of class ahull}

\item{add}{if TRUE the plot is added to the active graphic window.}

\item{do.shape}{if TRUE the \eqn{\alpha}-shape for the same sites and same level alpha is plotted.}

\item{wlines}{a string specifying what has to be plotted:
\itemize{
  \item "none": nothing is added to the plot;
  \item "vor": shows the Voronoi tesselation;
  \item "del": shows the Delanuay triangulation plot;
  \item "both": shows both of the above structures.
}}

\item{wpoints}{if true, the sites are added to the plot.}

\item{number}{if true, the plot shows the indeces of the sites between their locations.}

\item{col}{specifies in a vector the colors to be used for the different objects to be plotted and the
order is:
\itemize{
  \item col[1] -> color of the \eqn{\alpha}-hull;
  \item col[2] -> color of the \eqn{\alpha}-shape;
  \item col[3] -> color of the points;
  \item col[4] -> color of Delanuay triangulation;
  \item col[5] -> color of Voronoi tesselation
  \item col[6] -> color for the numbers
} 
Alternatively just one color can be provided.}

\item{xlim}{x axis limits.}

\item{ylim}{y axis limits.}

\item{...}{graphical arguments to be passes to methods (see \code{\link[graphics]{par}})}
}
\value{

}
\description{
Plot of an object "ahull".
}
\examples{
x = runif(10)
y = runif(10)
ahull.obj = ahull(x, y, alpha = 0.5)

plot.ahull(ahull.obj, wlines = "none", wpoints = TRUE)

}
\seealso{
\code{\link{ahull}}
}
