% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.delvor.cpp.R
\name{plot.delvor.cpp}
\alias{plot.delvor.cpp}
\title{Voronoi diagram/Delanuay tesselation plot}
\usage{
\method{plot}{delvor.cpp}(x, add = FALSE, wlines = c("both", "del",
  "vor"), wpoints = TRUE, number = FALSE, col = NULL, xlim = NULL,
  ylim = NULL, ...)
}
\arguments{
\item{x}{object of class del.vor}

\item{if}{TRUE add the plot to the current ative graphics windows}

\item{aaa}{}
}
\value{

}
\description{
Plot of an object "del.vor" with a "fake.tri" tri object
}
\examples{
x = runif(10)
y = runif(10)
my.delvor(x, y)

}
