% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppAlphahull-package}
\alias{RcppAlphahull}
\alias{RcppAlphahull-package}
\title{RcppAlphahull}
\description{
A package to compute Voronoi diagrams and Delanuay tesselations with the use of a
C++ library
}
\details{
RcppAlphahull is package aimed to compute Voronoi diagrams and Delanuay tesselations
for sets of points using the C++ package MyGAL (link...). The package is based on the
R/Fortran package alphahull which does the same job, but much slower.
}
\author{
Federico Airoldi
}
\keyword{Delanuay}
\keyword{Voronoi}
\keyword{diagram,}
\keyword{tesselation}
