% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashape.R
\name{ashape}
\alias{ashape}
\title{Alpha-shape computation}
\usage{
ashape(x, y = NULL, alpha)
}
\arguments{
\item{x}{coordinates of the sites}

\item{y}{coordinates of the sites}

\item{alpha}{}
}
\value{
A list with the following components: 
\describe{
   \item{edges}{a mesh describing the alpha shape}
   \item{length}{length of the alpha shape}
   \item{alpha}{the value of alpha for which the shape is computed}
   \item{alpha.extremes}{containes the indices of those sites that are alpha extremes}
}
}
\description{
Computes the alpha-shape for the given set of points
}
\examples{
x = runif(10)
y = runif(10)
a.shape = ashape(x, y, alpha = 0.8)
plot(a.shape)

}
