% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppAlphahull-package}
\alias{RcppAlphahull}
\alias{RcppAlphahull-package}
\title{RcppAlphahull}
\description{
RcppAlphahull is package aimed to compute Voronoi diagrams and Delanuay tesselations for sets of 
points using the C++ package \href{https://github.com/pvigier/FortuneAlgorithm}{MyGAL}. The package is
inspired by the R/Fortran package alphahull which does the same job, but much slower. Plot of some 
graphs are done by using this package functions.
}
\details{
Voronoi tesselation/Delanuay triangulation, \eqn{\alpha}-shape and \eqn{\alpha}-hull computation 
using a C++ library.
}
\references{
Edelsbrunner, H., Kirkpatrick, D. G., Seidel, R. 1983, 'On the Shape of a Set of 
Points in the Plane', \emph{IEEE Transactions on Information Theory}, vol. 29, no. 4, pp. 551-559.
}
\author{
Federico Airoldi \email{federico2.airoldi@mail.polimi.it}
}
\keyword{Delanuay}
\keyword{Voronoi}
\keyword{diagram,}
\keyword{tesselation}
