% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppAlphahull-package}
\alias{RcppAlphahull}
\alias{RcppAlphahull-package}
\title{RcppAlphahull}
\description{
Voronoi tesselation/Delanuay triangulation, \eqn{\alpha}-shape and \eqn{\alpha}-hull computation using 
a C++ library.
}
\details{
RcppAlphahull is package aimed to compute Voronoi diagrams and Delanuay tesselations for sets of 
points using the C++ package \href{https://github.com/pvigier/FortuneAlgorithm}{MyGAL}. The package is
inspired by the R/Fortran package alphahull which does the same job, but much slower. Plot of some 
graphs are done by using this package functions.
}
\author{
Federico Airoldi \email{federico2.airoldi@mail.polimi.it}
}
\keyword{Delanuay}
\keyword{Voronoi}
\keyword{diagram,}
\keyword{tesselation}
