% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RcppAlphahull-package}
\alias{RcppAlphahull}
\alias{RcppAlphahull-package}
\title{RcppAlphahull}
\description{
A package to compute Voronoi diagrams and Delanuay tesselations with the use of a
C++ library
}
\details{
RcppAlphahull is package aimed to compute Voronoi diagrams and Delanuay tesselations for huge sets of points with the aim of the C++ package MyGAL (link...).
The package is based on the R/Fortran package alphahull which does the same job, however, due to its slowliness (??) its use on sets with huge number of points was ...
Use four spaces when indenting paragraphs within the Description.
}
\keyword{Delanuay}
\keyword{Voronoi}
\keyword{diagram,}
\keyword{tesselation}
